etc/lsp/requirements.txt
pygls
lsprotocol
PyYAML
tree-sitter
pyinstaller
